+++ src/main/java/com/ddudu/infrastructure/persistence/repository/ddudu/DduduQueryRepository.java
@@ -1,4 +1,4 @@
-package com.ddudu.old.persistence.dao.todo;
+package com.ddudu.infrastructure.persistence.repository.ddudu;
 
 import com.ddudu.application.domain.goal.domain.enums.PrivacyType;
 import com.ddudu.infrastructure.persistence.entity.UserEntity;
@@ -7,7 +7,7 @@ import com.ddudu.old.todo.dto.response.TodoCompletionResponse;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public interface TodoDaoCustom {
+public interface DduduQueryRepository {
 
   List<TodoEntity> findTodosByDate(LocalDateTime startDate, LocalDateTime endDate, UserEntity user);
