+++ src/main/java/com/ddudu/infrastructure/persistence/repository/ddudu/DduduQueryRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.ddudu.old.persistence.dao.todo;
+package com.ddudu.infrastructure.persistence.repository.ddudu;
 
 import static com.ddudu.old.persistence.entity.QTodoEntity.todoEntity;
 
@@ -18,11 +18,11 @@ import java.util.List;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class TodoDaoImpl implements TodoDaoCustom {
+public class DduduQueryRepositoryImpl implements DduduQueryRepository {
 
   private final JPAQueryFactory jpaQueryFactory;
 
-  public TodoDaoImpl(EntityManager em) {
+  public DduduQueryRepositoryImpl(EntityManager em) {
     this.jpaQueryFactory = new JPAQueryFactory(em);
   }
