--- src/main/java/com/ddudu/old/persistence/dao/todo/TodoDao.java
@@ -1,8 +1,0 @@
-package com.ddudu.old.persistence.dao.todo;
-
-import com.ddudu.old.persistence.entity.TodoEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TodoDao extends JpaRepository<TodoEntity, Long>, TodoDaoCustom {
-
-}
